Add transcriber and converter classes for handling audio transcriptions
