+++ scinoephile/core/hanzi/hanzi_converter.py
@@ -1,0 +1,35 @@
+#  Copyright 2017-2025 Karl T Debiec. All rights reserved. This software may be modified
+#  and distributed under the terms of the BSD license. See the LICENSE file for details.
+"""Runnable for converting between Hanzi characters."""
+
+from __future__ import annotations
+
+from langchain_core.runnables import Runnable, RunnableConfig
+from opencc import OpenCC
+
+from scinoephile.audio import TranscribedSegment
+
+
+class HanziConverter(Runnable):
+    """Runnable for converting between Hanzi characters."""
+
+    def __init__(self, config: str = "s2hk"):
+        self.config = config
+        self.converter = OpenCC(config)
+
+    def invoke(
+        self,
+        input: list[TranscribedSegment],
+        config: RunnableConfig | None = None,
+        **kwargs,
+    ) -> list[TranscribedSegment]:
+        for segment in input:
+            segment.text = self.converter.convert(segment.text)
+            if segment.words:
+                i = 0
+                for word in segment.words:
+                    word_length = len(word.text)
+                    word.text = segment.text[i : i + word_length]
+                    i += word_length
+                joined = "".join([w.text for w in segment.words])
+        return input
