+++ test/data/mlamd/audio.py
@@ -6,82 +6,11 @@ from __future__ import annotations
 
 import time
 from pathlib import Path
-from typing import Any
 
-import whisper_timestamped as whisper
-from langchain_core.runnables import Runnable, RunnableConfig
-from opencc import OpenCC
-from pydantic import BaseModel, Field
-
-from scinoephile.audio import AudioSeries
-from scinoephile.common.file import get_temp_file_path
+from scinoephile.audio import AudioSeries, Transcriber
 from scinoephile.common.logs import set_logging_verbosity
+from scinoephile.core.hanzi import HanziConverter
 from scinoephile.testing import test_data_root
-
-
-class TimestampedWord(BaseModel):
-    """Single word within a transcribed segment."""
-
-    text: str = Field(..., description="Word's transcription.")
-    start: float = Field(..., description="Start time of word in seconds.")
-    end: float = Field(..., description="End time of word in seconds.")
-    confidence: float = Field(..., description="Confidence of transcription.")
-
-
-class TimestampedSegment(BaseModel):
-    """Transcribed segment."""
-
-    id: int = Field(..., description="Segment ID, usually sequential.")
-    seek: int = Field(..., description="Audio seek offset where segment starts.")
-    start: float = Field(..., description="Start time of segment in seconds.")
-    end: float = Field(..., description="End time of the segment in seconds.")
-    text: str = Field(..., description="Full transcription of segment.")
-    tokens: list[int] | None = Field(None, description="Token IDs for segment.")
-    temperature: float | None = Field(None, description="Sampling temperature.")
-    avg_logprob: float | None = Field(None, description="Average log-probability.")
-    compression_ratio: float | None = Field(None, description="Compression ratio.")
-    no_speech_prob: float | None = Field(None, description="Probability of no speech.")
-    words: list[TimestampedWord] | None = Field(None, description="Words in segments.")
-
-
-class Transcriber(Runnable):
-    def __init__(self, model_name: str = "khleeloo/whisper-large-v3-cantonese"):
-        self.model_name = model_name
-        self.model = whisper.load_model(model_name)
-
-    def invoke(
-        self, input: Any, config: RunnableConfig | None = None, **kwargs
-    ) -> list[dict]:
-        with get_temp_file_path(suffix=".wav") as temp_audio_path:
-            input.audio.export(temp_audio_path, format="wav")
-            result = whisper.transcribe(
-                self.model,
-                str(temp_audio_path),
-                language="yue",
-            )
-        return [TimestampedSegment(**s) for s in result["segments"]]
-
-
-class HanziConverter(Runnable):
-    def __init__(self, config: str = "s2hk"):
-        self.config = config
-        self.converter = OpenCC(config)
-
-    def invoke(
-        self,
-        input: list[TimestampedSegment],
-        config: RunnableConfig | None = None,
-        **kwargs,
-    ) -> list[TimestampedSegment]:
-        for segment in input:
-            segment.text = self.converter.convert(segment.text)
-            if segment.words:
-                for word in segment.words:
-                    word.text = self.converter.convert(word.text)
-            print(f"{len(segment.text)}, {len(segment.words)} words in segment")
-            print(f"|{segment.text}|{''.join([w.text for w in segment.words])}|")
-        return input
-
 
 if __name__ == "__main__":
     bluray_root = Path("/Volumes/Backup/Video/Disc")
@@ -98,8 +27,16 @@ if __name__ == "__main__":
     yue_hans = AudioSeries.load(test_output_dir / "yue-Hans_audio")
 
     # Runnables
+    # Code: Preprocess audio for transcription
+    # LLM: Transcribe audio to Cantonese
     transcriber = Transcriber("khleeloo/whisper-large-v3-cantonese")
+    # Code: Convert simplified Chinese to traditional
     hanzi_converter = HanziConverter("s2hk")
+    # Code: Convert transcriptions to subtitles
+    #   AudioSubtitle will need to store metadata such as words, timings, and confidence
+    # Code: Get sync groups between source subtitles and transcribed subtitles
+    #   Several source subtitles may map to each transcribed subtitle
+    # LLM:
 
     # Pipeline
     pipeline = transcriber | hanzi_converter
